Select the simplest type for macro expansions
All macro expansions should have compatible types, and if they do, we select the
"simplest" (smallest, requiring no additional casts) type as the type of the
macro constant. Uses of the macro may require casts to the required type, but
the macro definition itself will stay clear, simple, and compilable.